Turn on warning-is-error for sphinx build
The doc-migration spec requires all projects to treat warnings
as errors during sphinx build. This patches turns that function
on and resolves the error (detailed below) coming as a result
of it.

The code block in event_concept.rst file is currently made to
render in javascript which is an incorrect syntax for it.

Change-Id: Id7fee247b76ed7cfebb849305e0611bc966709be