Add py37 tox env and drop py35
With python3.7 jobs added to the Zuul jobs,
it makes sense to add this to the tox config
too so that developers can run python3.7 tests
as well.
Python 3.5 was the target runtime for the Rocky release.
The current target py3 runtime for Stein is Python 3.6,
so there is no reason to keep testing against the older version.

Change-Id: Icada08448d9bf3cc0f1d522d93b07d8789adad14