Removes temporary fix for doc generation
A temporary fix was added to get around a bug in how pbr handles
its autodoc_tree_index_modules setting. Since this bug is fixed we no
longer need the work around.

This change is needed since the workaround is now incompatible with the
latest version of Sphinx.

See also[1][2].

[1] https://review.opendev.org/191633
[2] https://review.opendev.org/121667

Change-Id: Idf9dc73d99b4233f49effe24aa047959b1a2ef2c