Fix pep8 gate
This updates the test-requirements to a more recent flake8-docstrings
version.  I also added an ignore for __init__ methods, this may be
something that we want to revisit but are not needed for now.

We also slip in a little change that renames whitelist to
allowlist. It is also a gate failure and thus is circular with
the pep8 fix: neither can be merged separately.

Change-Id: Id12f6e3b5d603531744d02f7bafab0b80696751e